requirements.txt
flask==3.0.2
requests==2.31.0
beautifulsoup4==4.12.3
python-dotenv==1.0.1
fake-useragent==1.4.0
urllib3==2.2.1
certifi==2024.2.2
pandas==2.1.0
twilio==8.3.0
apscheduler==3.10.1
selenium==4.10.0
webdriver-manager==3.8.6
requests-html==0.10.0
nltk==3.8.1
feedparser==6.0.11
gunicorn==21.2.0